# openapi-schema.yml
openapi: 3.0.2
info:
  title: ''
  version: ''
paths:
...
  /reservations/add/auth:
    post:
      operationId: createReservation
      description: ''
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reservation'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Reservation'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Reservation'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
          description: ''
      tags:
      - reservations
...
components:
  schemas:
...
    Reservation:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        car:
          type: object
          properties:
            id:
              type: integer
              readOnly: true
            image:
              type: array
              items:
                type: string
            model:
              type: string
              maxLength: 100
            doors:
              type: integer
            seats:
              type: integer
            luggage:
              type: integer
            transmission:
              enum:
              - a
              - m
              - t
              type: string
            airConditioning:
              type: boolean
            age:
              type: integer
            pricePerHour:
              type: string
              format: decimal
              multipleOf: 0.01
              maximum: 1000000
              minimum: -1000000
            fuelType:
              enum:
              - e
              - h
              - g
              - d
              - hy
              - l
              - c
              type: string
            builtIn:
              type: boolean
          required:
          - model
          - doors
          - seats
          - luggage
          - transmission
          - airConditioning
          - age
          - pricePerHour
          - fuelType
          readOnly: true
        pickUpTime:
          type: string
          format: date-time
        dropOffTime:
          type: string
          format: date-time
        car_id:
          type: integer
        user_id:
          type: integer
        pickUpLocation:
          type: string
          maxLength: 150
        dropOffLocation:
          type: string
          maxLength: 150
        status:
          type: string
          readOnly: true
        totalPrice:
          type: string
          format: decimal
          multipleOf: 0.01
          maximum: 1000000000000000000
          minimum: -1000000000000000000
          readOnly: true
        userId:
          type: string
          readOnly: true
      required:
      - car_id
      - user_id
      - pickUpLocation
      - dropOffLocation
...